feat: 면적 float값 허용하도록 pydantic 변경 / 작업 상태 ui 삭제

+++ BHL/services/database/app/models.py
@@ -43,7 +43,7 @@ class ExtractedPropertyInfo(BaseModel):
     property_type: Optional[str] = None
     memo: Optional[str] = None
     floor: Optional[int] = None
-    area: Optional[int] = None
+    area: Optional[float] = None
     premium: Optional[int] = None
     owner_property_memo: Optional[str] = None
     tenant_property_memo: Optional[str] = None
@@ -116,7 +116,7 @@ class PropertyInfo(BaseModel):
     property_type: Optional[str] = None
     memo: Optional[str] = None
     floor: Optional[int] = None
-    area: Optional[int] = None
+    area: Optional[float] = None
     premium: Optional[int] = None
     moving_date: Optional[datetime] = None
     owner_property_memo: Optional[str] = None
